backend/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
gtts
tortoise-tts
pydantic
numpy
